requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# AthenaFinal.py: 1,2
# controllers\analysis_controller\analysis_manager.py: 1
# controllers\analysis_controller\sentiment_view_analysis.py: 4
# controllers\analysis_controller\topic_view_analysis.py: 1
# controllers\download.py: 2
# controllers\uploadFile.py: 2
Flask == 0.12.2

# controllers\uploadFile.py: 3
Werkzeug == 0.12.2

# controllers\Feature_Extraction\NGramExtraction.py: 1
# controllers\Sentiment_Analysis\Sentiment_Identification.py: 1
afinn == 0.1

# controllers\visualization\donut_chart.py: 7,8,10
# controllers\visualization\scatterplot.py: 2,3,4
# test_bokeh.py: 25
# test_tsne.py: 8,9,10
bokeh == 0.12.5

# AthenaFinal.py: 9,10,11,12,13,14,15,16
# controllers\Candidate_Analysis\Candidate_Identification.py: 4
# controllers\DataCleaning\cleaning.py: 6,7,8
# controllers\Feature_Extraction\ngram_extractor.py: 5
# controllers\Sentiment_Analysis\Sentiment_Identification.py: 3
# controllers\Topic_Analysis\Find_Topic_Tweets.py: 2
# controllers\analysis_controller\analysis_manager.py: 2,3,4
# controllers\analysis_controller\sentiment_view_analysis.py: 1,3
# controllers\analysis_controller\topic_view_analysis.py: 2,3,4,6,7
# controllers\download.py: 4,5
# controllers\machine_learning\cleaning.py: 3,7
# controllers\machine_learning\machine_learning.py: 10,11
# controllers\uploadFile.py: 6
# controllers\visualization\donut_chart.py: 9
# test scripts\cleaning test.py: 2,8,9
# test scripts\sa_test.py: 1,2
# test scripts\test ml.py: 14,15,16
# test scripts\whole_db.py: 1,4
# testLDAvis.py: 2,3
crowdtruth == 1.0

# controllers\DataCleaning\Patterns.py: 2
emoji == 0.4.5

# controllers\Topic_Analysis\pyldavis_to_sklearn.py: 7
funcy == 1.8

# controllers\Topic_Analysis\Find_Topic_Tweets.py: 1
fuzzywuzzy == 0.15.0

# controllers\machine_learning\machine_learning.py: 14
# controllers\visualization\wordcloud_viz.py: 11,12
# test scripts\sa_test.py: 8
# test scripts\test ml.py: 9
matplotlib == 2.0.2

# controllers\Feature_Extraction\NGramExtraction.py: 2,3,4,5
# controllers\Feature_Extraction\ngram_extractor.py: 1,2
# controllers\Sentiment_Analysis\Sentiment_Identification.py: 2
# controllers\Topic_Analysis\Topic_Analysis.py: 7
nltk == 3.2.4

# controllers\Candidate_Analysis\Candidate_Identification.py: 7
# controllers\DataCleaning\cleaning.py: 12
# controllers\Sentiment_Analysis\Sentiment_Identification.py: 7
# controllers\Topic_Analysis\Topic_Analysis.py: 6
# controllers\machine_learning\cleaning.py: 6
# controllers\machine_learning\machine_learning.py: 16
# controllers\visualization\scatterplot.py: 1
# controllers\visualization\wordcloud_viz.py: 9
# test scripts\sa_test.py: 10
# test scripts\test ml.py: 13
# test_tsne.py: 7
numpy == 1.13.1

# controllers\visualization\wordcloud_viz.py: 8
omelette == 1.1.6

# controllers\Feature_Extraction\NGramExtraction.py: 12
openpyxl == 2.4.8

# controllers\visualization\wordcloud_viz.py: 16
palettable == 3.0.0

# DBModels\Lexicon.py: 4
# DBModels\Tweet.py: 4
# controllers\Candidate_Analysis\Candidate_Identification.py: 3
# controllers\DataCleaning\cleaning.py: 9
# controllers\DataCleaning\emojip.py: 2
# controllers\Feature_Extraction\NGramExtraction.py: 9
# controllers\Lexicon_Files\populate_lexicon_base.py: 1
# controllers\Lexicon_Files\save_fil_words.py: 1
# controllers\ManageTrainingData.py: 1
# controllers\Pickles\Pickle_Saver.py: 3
# controllers\download.py: 1
# controllers\init_athena.py: 2
# controllers\machine_learning\cleaning.py: 4
# controllers\uploadFile.py: 7
# controllers\visualization\donut_chart.py: 5
# controllers\visualization\wordcloud_viz.py: 18
# test scripts\Test.py: 2
# test scripts\cleaning test.py: 10
# test scripts\sa_test.py: 3
# test_bokeh.py: 26
pandas == 0.20.1

# controllers\visualization\wordcloud_viz.py: 8
pillow == 4.2.1

# controllers\visualization\wordcloud_viz.py: 8
pillow-simd == 4.2.1.post0

# controllers\visualization\wordcloud_viz.py: 10
prettyplotlib == 0.1.7

# controllers\Topic_Analysis\pyldavis_to_sklearn.py: 8
# testLDAvis.py: 1
pyldavis == 2.1.1

# DBModels\Data.py: 2
# DBModels\KBFile.py: 3
# DBModels\KB_Names.py: 2
# DBModels\Lexicon.py: 2
# DBModels\MongoDB_Manager.py: 2
# DBModels\Tweet.py: 2
# DBModels\Username.py: 2
pymodm == 0.4.0

# DBModels\Data.py: 1,4
# DBModels\KBFile.py: 2,5
# DBModels\KB_Names.py: 1,3
# DBModels\Lexicon.py: 1,3
# DBModels\MongoDB_Manager.py: 1,4
# DBModels\Tweet.py: 1,3,6
# DBModels\Username.py: 1,3,4,5
pymongo == 3.4.0

# AthenaFinal.py: 9,10,11,12,13,14,15,16
# controllers\Candidate_Analysis\Candidate_Identification.py: 4
# controllers\DataCleaning\cleaning.py: 6,7,8
# controllers\Feature_Extraction\ngram_extractor.py: 5
# controllers\Sentiment_Analysis\Sentiment_Identification.py: 3
# controllers\Topic_Analysis\Find_Topic_Tweets.py: 2
# controllers\analysis_controller\analysis_manager.py: 2,3,4
# controllers\analysis_controller\sentiment_view_analysis.py: 1,3
# controllers\analysis_controller\topic_view_analysis.py: 2,3,4,6,7
# controllers\download.py: 4,5
# controllers\machine_learning\cleaning.py: 3,7
# controllers\machine_learning\machine_learning.py: 10,11
# controllers\uploadFile.py: 6
# controllers\visualization\donut_chart.py: 9
# test scripts\cleaning test.py: 2,8,9
# test scripts\sa_test.py: 1,2
# test scripts\test ml.py: 14,15,16
# test scripts\whole_db.py: 1,4
# testLDAvis.py: 2,3
ricardo-crawler == 01.00.01

# controllers\Topic_Analysis\Topic_Analysis.py: 2,3,4,10
# controllers\machine_learning\machine_learning.py: 9
# test scripts\sa_test.py: 7
# test scripts\test ml.py: 11
# test_tsne.py: 1,2,4
sklearn == 0.0

# controllers\visualization\wordcloud_viz.py: 15
wordcloud == 1.3.1
